+++ target/generated-sources/r/net/margaritov/preference/colorpicker/BuildConfig.java
@@ -1,0 +1,5 @@
+package net.margaritov.preference.colorpicker;
+
+public final class BuildConfig {
+  public static final boolean DEBUG = true;
+}